feat: 쿼리 실행 클래스 분리
