+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -4,10 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.sql.DataSource;
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -17,30 +15,19 @@ public class JdbcTemplate {
     private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
 
     private final DataSource dataSource;
+    private final QueryTemplate queryTemplate;
 
     public JdbcTemplate(final DataSource dataSource) {
         this.dataSource = dataSource;
+        this.queryTemplate = new QueryTemplate(dataSource);
     }
 
     public void update(String sql, Object... parameters) {
-        try (
-                Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = conn.prepareStatement(sql)
-        ) {
-            settingParameters(pstmt, parameters);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
+        queryTemplate.query(sql, PreparedStatement::executeUpdate, parameters);
     }
 
     public <T> T queryForObject(String sql, RowMapper<T> rowMapper, Object... parameters) {
-        try (
-                Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = conn.prepareStatement(sql)
-        ) {
-            settingParameters(pstmt, parameters);
+        return queryTemplate.query(sql, (pstmt -> {
             try (ResultSet rs = pstmt.executeQuery()) {
                 log.debug("query : {}", sql);
                 if (rs.next()) {
@@ -48,17 +35,11 @@ public class JdbcTemplate {
                 }
                 throw new NoSuchElementException("조건에 맞는 데이터가 없습니다.");
             }
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
+        }), parameters);
     }
 
     public <T> List<T> query(String sql, RowMapper<T> rowMapper, Object... parameters) {
-        try (Connection conn = dataSource.getConnection();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
-
-            settingParameters(pstmt, parameters);
+        return queryTemplate.query(sql, (pstmt -> {
             try (ResultSet rs = pstmt.executeQuery()) {
                 log.debug("query : {}", sql);
                 List<T> results = new ArrayList<>();
@@ -67,15 +48,6 @@ public class JdbcTemplate {
                 }
                 return results;
             }
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void settingParameters(PreparedStatement pstmt, Object[] parameters) throws SQLException {
-        for (int index = 1; index < parameters.length + 1; index++) {
-            pstmt.setObject(index, parameters[index - 1]);
-        }
+        }), parameters);
     }
 }
