+++ jdbc/src/main/java/org/springframework/jdbc/core/QueryExecutor.java
@@ -1,0 +1,10 @@
+package org.springframework.jdbc.core;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface QueryExecutor<T> {
+
+    T execute(PreparedStatement pstmt) throws SQLException;
+}
