+++ jdbc/src/main/java/org/springframework/jdbc/core/QueryTemplate.java
@@ -1,0 +1,39 @@
+package org.springframework.jdbc.core;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public class QueryTemplate {
+
+    private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
+
+    private final DataSource dataSource;
+
+    public QueryTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public <T> T query(String sql, QueryExecutor<T> rExecution, Object... parameters) {
+        try (
+                Connection conn = dataSource.getConnection();
+                PreparedStatement pstmt = conn.prepareStatement(sql)
+        ) {
+            settingParameters(pstmt, parameters);
+            return rExecution.execute(pstmt);
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void settingParameters(PreparedStatement pstmt, Object[] parameters) throws SQLException {
+        for (int index = 1; index < parameters.length + 1; index++) {
+            pstmt.setObject(index, parameters[index - 1]);
+        }
+    }
+}
